YKI(Frontend): clean up unused/old wip parts
