+++ frontend/packages/yki/src/components/layouts/clerkHeader/ClerkHeader.tsx
@@ -2,14 +2,23 @@ import { AppBar, Toolbar } from '@mui/material';
 
 import { ClerkNavigationLinks } from 'components/layouts/clerkHeader/ClerkNavigationLinks';
 
-export const ClerkHeader = () => {
+export const ClerkHeader = (): JSX.Element => {
   return (
-    <AppBar className="header" position="static">
-      <Toolbar className="header__toolbar">
-        <div className="header__tabs">
-          <ClerkNavigationLinks />
-        </div>
-      </Toolbar>
-    </AppBar>
+    <>
+      <AppBar
+        className="header"
+        position="static"
+        sx={{
+          backgroundColor: 'white',
+          color: 'black',
+        }}
+      >
+        <Toolbar className="header__toolbar">
+          <div className="header__tabs">
+            <ClerkNavigationLinks />
+          </div>
+        </Toolbar>
+      </AppBar>
+    </>
   );
 };
