+++ frontend/packages/yki/src/redux/reducers/clerkOrganizer.ts
@@ -1,33 +1,22 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
 import { APIResponseStatus } from 'shared/enums';
 
-import {
-  ClerkOrganizer,
-  ClerkOrganizerFilters,
-} from 'interfaces/clerkOrganizer';
+import { ClerkOrganizer } from 'interfaces/clerkOrganizer';
 
 interface ClerkOrganizerState {
   organizers: Array<ClerkOrganizer>;
   status: APIResponseStatus;
-  filters: ClerkOrganizerFilters;
 }
 
 const initialState: ClerkOrganizerState = {
   organizers: [],
   status: APIResponseStatus.NotStarted,
-  filters: {},
 };
 
 const clerkOrganizersSlice = createSlice({
   name: 'clerkOrganizer',
   initialState,
   reducers: {
-    addClerkOrganizerFilter(
-      state,
-      action: PayloadAction<Partial<ClerkOrganizerFilters>>,
-    ) {
-      state.filters = { ...state.filters, ...action.payload };
-    },
     loadClerkOrganizers(state) {
       state.status = APIResponseStatus.InProgress;
     },
@@ -43,7 +32,6 @@ const clerkOrganizersSlice = createSlice({
 
 export const clerkOrganizersReducer = clerkOrganizersSlice.reducer;
 export const {
-  addClerkOrganizerFilter,
   loadClerkOrganizers,
   rejectClerkOrganizers,
   storeClerkOrganizers,
