+++ frontend/packages/yki/src/redux/selectors/clerkOrganizers.ts
@@ -1,18 +1,14 @@
 import { createSelector } from 'reselect';
 
 import { RootState } from 'configs/redux';
-import {
-  ClerkOrganizer,
-  ClerkOrganizerFilters,
-} from 'interfaces/clerkOrganizer';
+import { ClerkOrganizer } from 'interfaces/clerkOrganizer';
 
 export const clerkOrganizersSelector = (state: RootState) =>
   state.clerkOrganizer;
 
 export const selectFilteredClerkOrganizers = createSelector(
   (state: RootState) => state.clerkOrganizer.organizers,
-  (state: RootState) => state.clerkOrganizer.filters,
-  (organizers: ClerkOrganizer[], _filters: ClerkOrganizerFilters) => {
+  (organizers: ClerkOrganizer[]) => {
     // TODO: Add filters
     return organizers;
   },
